Pt_1_4: Full task
I wrote a program inviting the user to enter a number between 10 - 20.
If the entered value is less than 10,
the message "The number is less than the required
range" message is displayed and prompts the user to try again.
If the entered value is greater than 20,
The message "A number greater than the required range" is displayed,
and the user is prompted to try again.
Repeat until a value from the range of 10-20 is entered,
after which the "Thank you" message is displayed.