Pt_2_2: Full task
I created a list with the names of 4 TV shows and put them on separate
lines. I suggest that the user enter the name
of another program and the position at which
where it should be inserted into the list.
I display the list again, with all five programs at the new positions.