Pt_2_1: Full task
The program displays the names of five colors,
the program randomly chooses one and prompts the user to do the same.
If the user chooses the same color that the program
has chosen, the message "Fine!" is displayed; otherwise,
the answer hiding a hint to the right color.
If the colors do not match,
it prompts the user to try again;
if the user the user doesn't guess again,
 the same hint is displayed and the user is prompted to choose a color