Pt_2_5: Full task
I wrote a program that plays horoscope with the user.
The program must ask the user whether it is heads or tails.
If the user enters 0, he chooses heads,
1 - tails, any other number is the end of the game.
The program keeps a record of winnings and
of wins and losses after each round.
After three consecutive losses the game ends.