Pt_5_2: Full task
The user is asked how many records they want to add to the list,
then the user enters the records. After the entries are added,
the user is asked for the author and all books by the specified author
from the list are displayed.
Is asked for the author and all books by the specified author
from the list are displayed.If there are no in the list,
the corresponding message is displayed.
The books.csv file is used to solve the problem